requirements.txt
ultralytics>=8.0.0
opencv-python
torch
flask>=2.0.0
flask-cors>=3.0.0
pillow>=8.0.0
waitress>=2.0.0
numpy>=1.19.0